pyproject.toml
[tool.poetry]
name = "dotfive"
version = "0.1.0"
description = ""
authors = ["Stefan Andronic <stefan.and89@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.19.2"
SQLAlchemy = "^1.4.29"
pytest = "^6.2"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
